+++ lib/app/modules/notices/presentation/pages/feed_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:ziggle/app/di/locator.dart';
 import 'package:ziggle/app/modules/common/presentation/widgets/ziggle_app_bar.dart';
 import 'package:ziggle/app/modules/notices/presentation/bloc/notice_list_bloc.dart';
-import 'package:ziggle/app/modules/notices/presentation/widgets/notice_card.dart';
+import 'package:ziggle/app/modules/notices/presentation/widgets/list_layout.dart';
 import 'package:ziggle/app/values/palette.dart';
 
 class FeedPage extends StatelessWidget {
@@ -19,47 +19,8 @@ class FeedPage extends StatelessWidget {
       ),
       body: BlocProvider(
         create: (_) => sl<NoticeListBloc>()..add(const NoticeListEvent.load()),
-        child: const _Layout(),
+        child: const ListLayout(),
       ),
-    );
-  }
-}
-
-class _Layout extends StatelessWidget {
-  const _Layout();
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<NoticeListBloc, NoticeListState>(
-      builder: (context, state) {
-        return RefreshIndicator(
-          onRefresh: () async {
-            final bloc = context.read<NoticeListBloc>();
-            final blocker = bloc.stream.firstWhere((state) => !state.isLoading);
-            bloc.add(const NoticeListEvent.load());
-            await blocker;
-          },
-          child: state.showLoading
-              ? const Center(
-                  child: CircularProgressIndicator(),
-                )
-              : ListView.separated(
-                  padding:
-                      const EdgeInsets.symmetric(vertical: 20, horizontal: 16),
-                  itemBuilder: (context, index) {
-                    final notice = state.notices[index];
-                    return NoticeCard(
-                      onLike: () {},
-                      onPressed: () {},
-                      onShare: () {},
-                      notice: notice,
-                    );
-                  },
-                  separatorBuilder: (_, __) => const SizedBox(height: 15),
-                  itemCount: state.notices.length,
-                ),
-        );
-      },
     );
   }
 }
